switched power up, cards, character data, and battle data to enum IDs added rest view and model added treasure view and model blocked out shop model and view
added treasure and shop art

+++ Scripts/Cards/CardPile.cs
@@ -87,7 +87,7 @@ public partial class CardPile : Resource
 
         for (int i = 0; i < this.cards.Count; i++)
         {
-            data.Add("Card " + i.ToString(), cards[i].id);   
+            data.Add("Card " + i.ToString(),(int) cards[i].id);   
         }
         
         return data;
